chore: rename itest to test, move test_utils inside test::common
