+++ core/src/database/mod.rs
@@ -118,7 +118,7 @@ impl Database {
 
 #[cfg(test)]
 mod tests {
-    use crate::test_utils::*;
+    use crate::test::common::*;
     use crate::{config::BridgeConfig, database::Database};
 
     #[tokio::test]
