+++ core/src/database/tx_sender.rs
@@ -259,7 +259,7 @@ impl Database {
 #[cfg(test)]
 mod tests {
 
-    use crate::test_utils::*;
+    use crate::test::common::*;
 
     use super::*;
     use bitcoin::hashes::Hash;
