+++ core/src/operator.rs
@@ -874,7 +874,7 @@ impl Operator {
 #[cfg(test)]
 mod tests {
     use crate::operator::Operator;
-    use crate::test_utils::*;
+    use crate::test::common::*;
     use bitcoin::hashes::Hash;
     use bitcoin::Txid;
